builder/requirements.txt
python-pptx>=0.6.23
requests>=2.31.0
pydantic>=2.0.0
PyPDF2>=3.0.0
